requirements.txt
async-timeout==4.0.2
bcrypt==4.0.1
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==39.0.0
dm-tree==0.1.8
facenet-pytorch==2.5.3
Flask==2.2.5
idna==3.4
ifaddr==0.2.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mendel-development-tool==1.5.2
numpy==1.21.6
paramiko==3.0.0
Pillow==9.5.0
pycparser==2.21
PyNaCl==1.5.0
requests==2.30.0
tensorflow-addons==0.19.0
tensorflow-model-optimization==0.7.3
torch==1.13.1
torchvision==0.14.1
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==2.0.2
Werkzeug==2.2.3
zeroconf==0.47.1
zipp==3.15.0
